Create README.md
Hangman is a guessing game for two or more players. 
One player thinks of a word, and the other tries to guess 
it by suggesting letters within a certain number of guesses and in it player having some life line
depending on level and players can take hints also how much he want.